+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+- Added types: timestamp, timeuuid and counter
+- Fixed a bug where it would crash on selecting and inserting a null field
+
 ## 0.0.2
 
 - *BREAKING CHANGE*: `PageRows` implements `Page<Row>` from `package:page`.